Add experiments table footer and move select branches action
