+++ webview/src/experiments/components/table/commitsAndBranches/AddAndRemoveBranches.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { useSelector } from 'react-redux'
 import styles from './styles.module.scss'
-import { ExperimentsState } from '../../../../store'
-import { selectBranches } from '../../../../util/messages'
+import { ExperimentsState } from '../../../store'
+import { selectBranches } from '../../../util/messages'
 
 export const AddAndRemoveBranches: React.FC = () => {
   const { hasBranchesToSelect } = useSelector(
