+++ webview/src/experiments/components/table/commitsAndBranches/styles.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../shared/variables';
+@import '../../../../shared/variables';
 
 .commitsAndBranchesNav {
   text-align: left;
